r925	のコミットもれ
git-svn-id: http://aipo.googlecode.com/svn/trunk@930 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ portlets/webmail/src/main/java/com/aimluck/eip/webmail/WebMailFilterFormData.java
@@ -499,7 +499,7 @@ public class WebMailFilterFormData extends ALAbstractFormData {
       for (EipTMailFolder folder : mailFolders) {
         WebMailFolderResultData rd = new WebMailFolderResultData();
         rd.initField();
-        rd.setFolderId(folder.getFolderId().longValue());
+        rd.setFolderId(folder.getFolderId());
         rd.setFolderName(folder.getFolderName());
         folderList.add(rd);
       }
